requirements.txt
alembic==1.12.0
autopep8==1.6.0
blinker==1.6.3
click==8.1.3
colorama==0.4.6
dnspython==2.4.2
dominate==2.8.0
email-validator==2.1.0.post1
flake8==4.0.1
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==20.1.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.950
mypy-extensions==0.4.3
numpy==1.26.1
pyarrow==13.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylance==0.8.7
python-dotenv==0.20.0
SQLAlchemy==2.0.22
style==1.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
update==0.0.1
visitor==0.1.3
Werkzeug==2.3.0
WTForms==3.1.0
